requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
filelock==3.12.2
imageio==2.31.1
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
packaging==23.1
Pillow==10.0.0
pytz==2023.3
PyWavelets==1.4.1
scikit-image==0.21.0
scipy==1.11.1
sqlparse==0.4.4
sympy==1.12
tifffile==2023.7.18
torch==2.0.1
typing_extensions==4.7.1
